consolidate shared domainPermission code
